Move VerificationMethodButton to core
