+++ front/app/modules/commercial/id_auth0/components/Auth0Button.tsx
@@ -7,7 +7,7 @@ import {
   IDAuth0Method,
 } from 'services/verificationMethods';
 // components
-import VerificationMethodButton from 'modules/commercial/verification/citizen/components/VerificationMethodButton';
+import VerificationMethodButton from 'containers/App/VerificationModal/VerificationMethodButton';
 
 // i18n
 import T from 'components/T';
