+++ front/app/modules/commercial/id_bosa_fas/components/BosaFasButton.tsx
@@ -7,7 +7,7 @@ import { getJwt } from 'utils/auth/jwt';
 import { TVerificationMethod } from 'services/verificationMethods';
 
 // components
-import VerificationMethodButton from 'modules/commercial/verification/citizen/components/VerificationMethodButton';
+import VerificationMethodButton from 'containers/App/VerificationModal/VerificationMethodButton';
 
 // i18n
 import { FormattedMessage } from 'utils/cl-intl';
